clean up home file
