+++ src/utils/interfaces.tsx
@@ -1,0 +1,7 @@
+export interface UserDoc {
+  color: string;
+  email: string;
+  status: string;
+  uid: string;
+  username: string;
+}
